Work towards balancing power cost

+++ src/main/java/com/builtbroken/mffs/api/modules/IFortronCost.java
@@ -8,5 +8,5 @@ package com.builtbroken.mffs.api.modules;
  */
 public interface IFortronCost
 {
-    float getFortronCost(float paramFloat);
+    int getFortronCost(float paramFloat);
 }
